Implement base form-to-attribute translation
Creates a new namespace, Metadata::Ingest::Translators, for housing all
translation objects.  Adds a base interface and a very simple
delegation-style translator to take ingest form data and simply pass it
through to an object based on map rules.

Should be generally useable, and for OD's purposes we could subclass
and just hook into `store_attribute_value` if we need specialized logic
prior to sending the value to the object.